Scouting now looks for enemies
Enemy location is set once and never changed (definitely an issue for further down the line).
Fixed an issue with storing IDs on the shared array. IDs will act up if they are greater than 2^16 -1 (Once again to be fixed in the future)
Added records of how many miners and soldiers have been created as well as a template for future types of robots.